port avx2 implementation over to vortex-compute
