+++ vortex-compute/Cargo.toml
@@ -29,8 +29,10 @@ vortex-vector = { workspace = true }
 arrow-array = { workspace = true, optional = true }
 arrow-buffer = { workspace = true, optional = true }
 arrow-schema = { workspace = true, optional = true }
+log = { workspace = true }
 multiversion = { workspace = true }
 num-traits = { workspace = true }
+paste = { workspace = true }
 
 [features]
 default = ["arrow"]
